+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 - Update `pkg/network/anonymity`: delete multi addresses in the database hashes
 - Update `cmd/hidden_lake`: delete prefix 'go-peer/' from services
 - Update `pkg/crypto/keybuilder`: from bits -> to iter 
+- Update `cmd/hidden_lake/service`: delete check len(connections) from HandleNetworkRequestAPI 
 
 ### BUG FIXES